Standard File Manger
Fixed reading from built in file manager with a hack. I still have to
tweak it to work with added filemanager as I broke that. Doesn't work
with google drive. I also changed it so it doesn't crash if I change a
double into a string in the excel it will print a system.out saying data
not in right format. I gotta find a a way to stop the file from
importing if so. For memory management to stop a crash should I limit
the amount of data sets for the prototype? I could probably set a hard
limit that can be lifted based of tests when we have a working 3d point
graph running.